requirements.txt
google-cloud-aiplatform==1.60.*
flask==3.0.*
google-cloud-bigquery==3.25.*